# openapi.yaml
openapi: 3.1.0
info:
  title: Wander the Atlas — Geo Quiz API
  version: 1.0.0
  summary: Content, profile and session API for the kids' geography quiz.
  description: |
    Contract for the backend behind the `frontend/` app (TanStack Start + React).

    **How this maps to the current frontend.** v1 of the app is local-first: the
    question bank is bundled at build time (`frontend/src/data/`) and profiles live in
    `localStorage` (`frontend/src/lib/storage.ts`). This contract describes the same
    data as an HTTP surface so the bundled data can be replaced by fetched content
    without reshaping the client, and so profiles/sessions can move server-side later
    without a schema migration.

    Three groups of operations, deliberately separable:

    - **Content** (`/entities`, `/questions`, `/bundles`, `/geometry`,
      `/elevation-profiles`, `/superlative-axes`) — read-only, build-time generated,
      cacheable, no user data. This is the group worth implementing first; a static
      file host satisfies it.
    - **Profiles** (`/profiles`) — optional in v1. The app works with `localStorage`
      alone; these endpoints exist for the v1.5 export/import story and any later
      server accounts. Implementing them means storing children's data, so treat
      adoption as a product decision, not a default.
    - **Sessions** (`/sessions`) — server-side question selection and grading. The
...

...
    are derived for display and never stored:

    ```
...
    ```

...

...

...

...

  license:
...
servers:
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Create a grown-up account
      description: |
        One account per grown-up; profiles hang off it. The password is hashed with
        PBKDF2-HMAC-SHA256 and a per-account salt before it is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "409":
          description: That username is already registered.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        default:
          $ref: "#/components/responses/Problem"
  /auth/token:
    post:
      operationId: issueToken
      tags: [Auth]
      summary: Exchange credentials for a bearer token
      description: |
        An unknown username and a wrong password fail identically — which half was
        wrong is not the caller's business.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: A bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /auth/me:
    get:
      operationId: getCurrentAccount
      tags: [Auth]
      security:
        - bearerAuth: []
      summary: The signed-in account
      responses:
        "200":
          description: The account the token belongs to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      security:
        - bearerAuth: []
      summary: Revoke the presented token
      description: Tokens are opaque and server-side, so signing out really revokes.
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The opaque token from `POST /auth/token`, sent as
        `Authorization: Bearer <token>`. Not a JWT: nothing here needs a token a
        third party can verify offline, and an opaque token can be revoked.
...
  responses:
    Problem:
      description: Unexpected error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, malformed or expired bearer token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ValidationFailed:
      description: The request body failed validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
          example: Profile not found
        status:
          type: integer
          example: 404
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          description: Field-level validation failures.
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
                example: /name
              message:
                type: string
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 254
          example: grownup@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
          description: Hashed server-side; never returned.
    TokenRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
          description: Opaque. Only its digest is stored server-side.
        tokenType:
          type: string
          enum: [bearer]
        expiresIn:
          type: integer
          description: Seconds until the token stops resolving.
          example: 43200
    Account:
      type: object
      description: A grown-up's account. Profiles belong to exactly one.
      required: [id, username, createdAt]
      properties:
        id:
          type: string
          example: acct-demo
        username:
          type: string
        createdAt:
          type: string
          format: date-time